chore: rename folders and fix import paths
