+++ server/controllers/ProductController.js
@@ -1,5 +1,5 @@
 // controllers/productController.js
-import ProductService from '../services/productService.js';
+import ProductService from '../services/ProductService.js';
 
 export const getProducts = async (req, res, next) => {
   try {
